generate epoch_sync_info on header sync
