+++ integration-tests/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod genesis_helpers;
+pub mod nearcore_utils;
 pub mod node;
 pub mod runtime_utils;
 pub mod test_helpers;
